Removed legacy views for inspecting system state in debugger window.
