--- Atari2600/Debugger/MemoryView.xib
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="32700.99.1234" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
-    <dependencies>
-        <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="22690"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="MemoryViewController" customModule="Atari2600" customModuleProvider="target">
-            <connections>
-                <outlet property="view" destination="h9o-Ds-ArK" id="9n9-nO-zwD"/>
-            </connections>
-        </customObject>
-        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
-        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <textField focusRingType="none" horizontalHuggingPriority="251" verticalHuggingPriority="750" id="h9o-Ds-ArK" userLabel="Memory" customClass="DebuggerValueLabel" customModule="Atari2600" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="324" height="104"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMinY="YES"/>
-            <textFieldCell key="cell" lineBreakMode="clipping" id="XAo-kz-Gra">
-                <font key="font" size="11" name="SFMono-Regular"/>
-                <string key="title">0a 2a cf 18 4d de 46 15 63 04 e4 d2 dd 2e d0 c1
-c2 57 05 a3 c1 34 a3 fa a2 ca 26 dc d8 48 a9 1d
-77 66 cc 46 a2 be 24 6c ac 79 d3 76 08 94 c2 cd
-20 97 e5 e4 91 57 28 dd 1a 29 b7 15 6c 0f 83 e7
-cc 9a fc 48 6c a7 e4 fc f4 eb 42 36 70 61 df 20
-d6 11 5a dd c3 65 71 82 48 03 04 dc 18 ec c3 4f
-c1 19 d0 48 a1 33 57 db 78 f4 17 4f 67 f8 ec 2a
-e9 60 5e e7 66 8a 6d d4 80 df 54 45 1e dc 71 d4</string>
-                <color key="textColor" name="disabledControlTextColor" catalog="System" colorSpace="catalog"/>
-                <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-            </textFieldCell>
-            <point key="canvasLocation" x="34" y="-216"/>
-        </textField>
-    </objects>
-</document>
