--- Atari2600/Debugger/TimerView.xib
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="21701" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="21701"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="TimerViewController" customModule="Atari2600" customModuleProvider="target">
-            <connections>
-                <outlet property="cyclesLabel" destination="FTo-6T-do6" id="tGB-SV-zGm"/>
-                <outlet property="intervalsLabel" destination="0HI-x0-xP1" id="vNj-pU-83Z"/>
-                <outlet property="view" destination="Fcb-Hh-WDb" id="dTu-7b-4mT"/>
-            </connections>
-        </customObject>
-        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
-        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <customView id="Fcb-Hh-WDb">
-            <rect key="frame" x="0.0" y="0.0" width="260" height="31"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMinY="YES"/>
-            <subviews>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="0HI-x0-xP1" userLabel="Cycles value" customClass="DebuggerValueLabel" customModule="Atari2600" customModuleProvider="target">
-                    <rect key="frame" x="128" y="18" width="59" height="13"/>
-                    <textFieldCell key="cell" lineBreakMode="clipping" title="165 (a5)" id="Cls-GJ-dg0">
-                        <font key="font" size="11" name="SFMono-Regular"/>
-                        <color key="textColor" name="disabledControlTextColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="FTo-6T-do6" userLabel="Intervals value" customClass="DebuggerValueLabel" customModule="Atari2600" customModuleProvider="target">
-                    <rect key="frame" x="129" y="0.0" width="52" height="13"/>
-                    <textFieldCell key="cell" lineBreakMode="clipping" title="1365 /8" id="0fD-l9-lZf">
-                        <font key="font" size="11" name="SFMono-Regular"/>
-                        <color key="textColor" name="disabledControlTextColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="DN9-NZ-b68">
-                    <rect key="frame" x="71" y="18" width="53" height="14"/>
-                    <textFieldCell key="cell" lineBreakMode="clipping" title="Intervals:" id="K06-OW-32x">
-                        <font key="font" metaFont="system" size="11"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="0lG-yQ-4xT">
-                    <rect key="frame" x="81" y="0.0" width="43" height="14"/>
-                    <textFieldCell key="cell" lineBreakMode="clipping" title="Cycles:" id="ZlF-7F-I67">
-                        <font key="font" metaFont="system" size="11"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-            </subviews>
-            <constraints>
-                <constraint firstItem="FTo-6T-do6" firstAttribute="leading" secondItem="0lG-yQ-4xT" secondAttribute="trailing" constant="9" id="Bb2-p7-0Kg"/>
-                <constraint firstItem="0HI-x0-xP1" firstAttribute="top" secondItem="Fcb-Hh-WDb" secondAttribute="top" id="Fro-fH-yLh"/>
-                <constraint firstItem="0lG-yQ-4xT" firstAttribute="top" secondItem="DN9-NZ-b68" secondAttribute="bottom" constant="4" id="J06-5a-KGx"/>
-                <constraint firstItem="0HI-x0-xP1" firstAttribute="leading" secondItem="Fcb-Hh-WDb" secondAttribute="centerX" id="PHf-Td-RSb"/>
-                <constraint firstItem="0lG-yQ-4xT" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="Fcb-Hh-WDb" secondAttribute="leading" id="Rad-UR-zaL"/>
-                <constraint firstItem="0lG-yQ-4xT" firstAttribute="firstBaseline" secondItem="FTo-6T-do6" secondAttribute="firstBaseline" id="Rg3-q5-dql"/>
-                <constraint firstItem="0lG-yQ-4xT" firstAttribute="trailing" secondItem="DN9-NZ-b68" secondAttribute="trailing" id="Sbk-cZ-Fbb"/>
-                <constraint firstItem="DN9-NZ-b68" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="Fcb-Hh-WDb" secondAttribute="leading" id="XlA-Ru-YIw"/>
-                <constraint firstAttribute="trailing" relation="greaterThanOrEqual" secondItem="FTo-6T-do6" secondAttribute="trailing" id="YWo-xQ-urt"/>
-                <constraint firstItem="0HI-x0-xP1" firstAttribute="leading" secondItem="DN9-NZ-b68" secondAttribute="trailing" constant="8" symbolic="YES" id="dUa-nX-pjL"/>
-                <constraint firstItem="DN9-NZ-b68" firstAttribute="firstBaseline" secondItem="0HI-x0-xP1" secondAttribute="firstBaseline" id="dd2-AX-O0X"/>
-                <constraint firstAttribute="bottom" secondItem="FTo-6T-do6" secondAttribute="bottom" id="eEs-2w-kcW"/>
-                <constraint firstAttribute="trailing" relation="greaterThanOrEqual" secondItem="0HI-x0-xP1" secondAttribute="trailing" id="rfD-zP-XAK"/>
-            </constraints>
-            <point key="canvasLocation" x="112" y="-120"/>
-        </customView>
-    </objects>
-</document>
